Add JavaDoc
Currently most methods and class headers are not added.

This may cause confusion on what each class represents and what
each method do.

Add JavaDoc to most non-trivial classes and methods.

It is done so that others who look at this code and want to improve it,
knowns what each class or method do at a glance.