Add title bar of the chatbot window
The current window lacks a title bar.

This needs to be change so that users know
what is the chat bot they are interacting with.

Add a function to display the name of the app
in the title bar.

It is done this way as this is how the stage
is set up.